Add requirements.txt and update README.rst
It has become apparent that acccess to specific libraries is necessary
to read the excel files (notably, xlrd). Not a problem, but makes more
sense to include this as a requirements.txt and use pip to install
everything.

Still leaving out Gooey and wxWidgets as wxPython isn't available
through pip, and not everyone will want/need a GUI.

Signed-off-by: Rod Manning <rod.manning@taslytic.com>