WIP Commit to Complete sort_data() Function
Changed the purpose of this function to instead of process *all* of the
data, to simply re-factor it into individual df's for the various sort
criteria.

Another function, e.g. group_data(), will then do the work to apply the
timegrouper and sum the resulting timeseries.

Tests are mostly complete, but some additional work is needed to test
that the outcome from the function is as expected (it looks good, just
need to write test values for this).

Signed-off-by: Rod Manning <rod.manning@taslytic.com>