Bugfix: Show numbers at end of period
Implemeted a basic functionality to control the frequency/periods for
which resource availability is shown. The unittest for this resulted in
a bug being noted -- the label for when the cumsum() changed reflected
the *start* of the period, so (for short periods, anyhow) it looked like
the change was happening a few days before it actually did.

Changed this by adding the argument 'label="right"' in the pd.Grouper()
function. This makes the ts label reflect the *end* of the period.

(Potential) issue with this is that a change only now shows up the
day *after* it comes into effect -- e.g. if a change occurs on
01-01-2017, then a period *ending* on 01-01-2017 will not show it, but
the next period (say, 04-01-2017) *will* show it.

This is likely to be conservative for most resource planning. (Hard to
see how to build any additional logic into this, anyhow.)

Signed-off-by: Rod Manning <rod.manning@taslytic.com>