Add plot_data() function
This function takes a dict containing name/df pairs, plots the df's
using Seaborn, and then returns the result as a dict of name/StringIO
data.

Also added a helper function, __style_plot() that is used to apply
standard styles and markup to the plots.

Still have to write the unittests for this (once I work out how to test
a StringIO as an output ...?)

Signed-off-by: Rod Manning <rod.manning@taslytic.com>